requirements.txt
customtkinter==5.2.0
opencv-python==4.8.1.78
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2